git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Trunk@47774 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/com/pennant/backend/model/amtmasters/OwnerShipType.java
@@ -100,7 +100,7 @@ public class OwnerShipType implements java.io.Serializable, Entity {
 		return ownerShipTypeId;
 	}
 	
-	public void setOwnerShipTypeId(int ownerShipTypeId) {
+	public void setOwnerShipTypeId(long ownerShipTypeId) {
 		this.ownerShipTypeId = ownerShipTypeId;
 	}
